+++ gensched/__main__.py
@@ -1,0 +1,5 @@
+import sys
+
+from .cli import main
+
+sys.exit(main())
